Fixed random number's problem

+++ 3bit/src/it/unipv/sfw/trebit/model/game/games/IGames.java
@@ -1,5 +1,5 @@
 package it.unipv.sfw.trebit.model.game.games;
 
 public interface IGames {
-	double[] turn(int extra);
+	int[] turn(int extra);
 }
